requirements.txt
albumentations==0.4.6
certifi==2020.6.20
chardet==3.0.4
coloredlogs==14.0
cycler==0.10.0
decorator==4.4.2
docutils==0.16
dominate==2.5.1
future==0.18.2
humanfriendly==8.2
idna==2.10
imageio==2.9.0
imgaug==0.4.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
Kivy==1.11.1
kivy-deps.glew==0.2.0
kivy-deps.gstreamer==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
kiwisolver==1.2.0
matplotlib==3.3.0
networkx==2.4
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.1.0
Pillow==7.2.0
pygame==1.9.6
Pygments==2.6.1
pyparsing==2.4.7
pypiwin32==223
pyreadline==2.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.1.0
Shapely==1.7.0
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.7.24
torch==1.5.0+cpu
torchfile==0.1.0
torchvision==0.6.0+cpu
tornado==6.0.4
tqdm==4.48.2
urllib3==1.25.10
visdom==0.1.8.9
websocket-client==0.57.0
